Change start script to use babel-node
Babel-node is a shell command that behaves like Node but transpiles
files before they're passed to Node. This still currently breaks, however.
That's because we haven't told Babel what targets to build for, or what
changes to make to the code. Right now, Babel is taking our code in, and
spitting it back out without modifying it.

babel-node is designed for use in development, so I'm also changing the
name of the script to reflect that.
https://babeljs.io/docs/en/babel-node